Move errors into a common file
